moved dummyData and constant to contants folder
